+++ apps/fluent-tester/src/FluentTester/TestComponents/Button/deprecated/ButtonIconTest.tsx
@@ -3,7 +3,7 @@ import { Button, Text } from '@fluentui/react-native';
 import { Stack } from '@fluentui-react-native/stack';
 import * as React from 'react';
 import { Platform, View } from 'react-native';
-import { stackStyle } from '../Common/styles';
+import { stackStyle } from '../../Common/styles';
 import { SvgIconProps } from '@fluentui-react-native/icon';
 import TestSvg from './test.svg';
 
@@ -12,7 +12,7 @@ const CustomizedIconButton = Button.customize({
   content: { style: { marginStart: 5 } },
 });
 
-export const ButtonIconTest: React.FunctionComponent = () => {
+export const ButtonIconTest_deprecated: React.FunctionComponent = () => {
   const testImage = require('./icon_24x24.png');
 
   const svgProps: SvgIconProps = {
